refactor(tv): use dummy items from core
