+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/addserver/AddServerScreen.kt
@@ -50,7 +50,7 @@ import dev.jdtech.jellyfin.setup.presentation.addserver.AddServerAction
 import dev.jdtech.jellyfin.setup.presentation.addserver.AddServerEvent
 import dev.jdtech.jellyfin.setup.presentation.addserver.AddServerState
 import dev.jdtech.jellyfin.setup.presentation.addserver.AddServerViewModel
-import dev.jdtech.jellyfin.ui.dummy.dummyDiscoveredServer
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyDiscoveredServer
 import dev.jdtech.jellyfin.utils.ObserveAsEvents
 
 @Composable
