+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/components/DiscoveredServerItem.kt
@@ -25,7 +25,7 @@ import androidx.tv.material3.Surface
 import androidx.tv.material3.Text
 import dev.jdtech.jellyfin.presentation.theme.FindroidTheme
 import dev.jdtech.jellyfin.presentation.theme.spacings
-import dev.jdtech.jellyfin.ui.dummy.dummyServer
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyServer
 import dev.jdtech.jellyfin.core.R as CoreR
 
 @Composable
