+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/components/ServerItem.kt
@@ -24,7 +24,7 @@ import androidx.tv.material3.Surface
 import androidx.tv.material3.Text
 import dev.jdtech.jellyfin.presentation.theme.FindroidTheme
 import dev.jdtech.jellyfin.presentation.theme.spacings
-import dev.jdtech.jellyfin.ui.dummy.dummyDiscoveredServer
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyDiscoveredServer
 
 @Composable
 fun ServerItem(
