+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/components/UserItem.kt
@@ -29,7 +29,7 @@ import coil.request.ImageRequest
 import dev.jdtech.jellyfin.models.User
 import dev.jdtech.jellyfin.presentation.theme.FindroidTheme
 import dev.jdtech.jellyfin.presentation.theme.spacings
-import dev.jdtech.jellyfin.ui.dummy.dummyUser
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyUser
 import org.jellyfin.sdk.model.api.ImageType
 import dev.jdtech.jellyfin.core.R as CoreR
