+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/servers/ServerScreen.kt
@@ -35,8 +35,8 @@ import dev.jdtech.jellyfin.setup.presentation.servers.ServersAction
 import dev.jdtech.jellyfin.setup.presentation.servers.ServersEvent
 import dev.jdtech.jellyfin.setup.presentation.servers.ServersState
 import dev.jdtech.jellyfin.setup.presentation.servers.ServersViewModel
-import dev.jdtech.jellyfin.ui.dummy.dummyDiscoveredServer
-import dev.jdtech.jellyfin.ui.dummy.dummyServer
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyDiscoveredServer
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyServer
 import dev.jdtech.jellyfin.utils.ObserveAsEvents
 import java.util.UUID
