+++ app/tv/src/main/java/dev/jdtech/jellyfin/presentation/setup/users/UsersScreen.kt
@@ -35,7 +35,7 @@ import dev.jdtech.jellyfin.setup.presentation.users.UsersAction
 import dev.jdtech.jellyfin.setup.presentation.users.UsersEvent
 import dev.jdtech.jellyfin.setup.presentation.users.UsersState
 import dev.jdtech.jellyfin.setup.presentation.users.UsersViewModel
-import dev.jdtech.jellyfin.ui.dummy.dummyUsers
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyUsers
 import dev.jdtech.jellyfin.utils.ObserveAsEvents
 import dev.jdtech.jellyfin.setup.R as SetupR
