+++ app/tv/src/main/java/dev/jdtech/jellyfin/ui/SeasonScreen.kt
@@ -26,7 +26,7 @@ import dev.jdtech.jellyfin.models.PlayerItem
 import dev.jdtech.jellyfin.presentation.theme.FindroidTheme
 import dev.jdtech.jellyfin.presentation.theme.spacings
 import dev.jdtech.jellyfin.ui.components.EpisodeCard
-import dev.jdtech.jellyfin.ui.dummy.dummyEpisodeItems
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyEpisodeItems
 import dev.jdtech.jellyfin.utils.ObserveAsEvents
 import dev.jdtech.jellyfin.viewmodels.PlayerItemsEvent
 import dev.jdtech.jellyfin.viewmodels.PlayerViewModel
