+++ app/tv/src/main/java/dev/jdtech/jellyfin/ui/components/EpisodeCard.kt
@@ -29,7 +29,7 @@ import androidx.tv.material3.Text
 import dev.jdtech.jellyfin.models.FindroidEpisode
 import dev.jdtech.jellyfin.presentation.theme.FindroidTheme
 import dev.jdtech.jellyfin.presentation.theme.spacings
-import dev.jdtech.jellyfin.ui.dummy.dummyEpisode
+import dev.jdtech.jellyfin.core.presentation.dummy.dummyEpisode
 
 @Composable
 fun EpisodeCard(
